pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "race-ticker"
version = "0.1.0"
description = "LED race ticker display system"
requires-python = ">=3.8"
dependencies = [
    "flask>=2.0.0",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]
